Deal with multiple test-implementations.
As soon as we allow multiple test-implementations we get
into a mess, as Mauve regards an ID as unique and that is
based upon the test-definition not the implementing method

We want to allow:


* HTTPS test to succeed.
* SSL-check to fail.

Which means multiple tests of type "https" will have different
IDs.  Force this by adding on the class of the implementation.