Updated to use types for test-returns.
This commit introduces a major change in custodian.  In the past
any test had either two results:

* return false - The test failed.
* return true  - The test passed.

We're now using an enum, more or less, such that a test may return a
"skipped" result which will neither raise nor clear any alert(s).

This is useful in its own right but is being introduced specifically
to allow SSL-certificate tests to avoid raising and clearing outside
working days/hours.

This closes #10328.
This fixes #10328.