Ensure that we correctly parse bogus macro-definitions.
We've always had an implicit rule in macro-definitions, that they
end with a period.  This meant that the first line is valid:

        FOO is bar.vm.bytemark.co.uk.

However we'd expect this to fail:

        FOO is bar.vm.bytemark.co.uk

A similar issue would arise if a macro-definition involved more than
one host, only the first would be valid.

We've fixed this now, such that the trailing period is optional.