Removed references and support for beanstalkd.
The beanstalkd queue used to be used in the past, and we later
added support for Redis via a simple abstraction layer.  But now
we've no longer tested and used beanstalkd for over a year, and
the client-libraries are no longer available as native Debian
packages.

With that in mind we've excised the code, although left the
abstraction-class in-place.