Fallback to using `openssl` if we can't get certificates.
Since the ruby version available to wheezy doesn't support TLS 1.2
fetching the certificate from remote HTTPS servers will fail, if
that is all that is available.

If we hit that condition, and only that one, we'll fall back to
invoking `openssl` natively.  This will allow us to monitor
expiration-time for remote SSL certificates, but the downside is
that we no longr receive the bundle that the remote server might
send - so we cannot validate the signature chain.

This closes #2.